Implement <selectedoption> for StylableSelect
This patch adds support for the <selectedoption> element to <select>.
The <selectedoption> element replaces its innerHTML with the ancestor
<select>'s selected <option>s innerHTML whenever the selected <option>
changes.

This patch implements it using a SynchronousMutationObserver with an
IsDescendantOf check, which might be slow. It could be worth combining
it with a SynchronousMutationObserver in HTMLSelectElement like
HTMLSelectListElement does, but this element will likely be specced and
shipped after StylableSelect or not get shipped at all, so I think at
least for now it is worth having a separate SynchronousMutationObserver
like this.

Bug: 1511354
Change-Id: I38be4657e11dd766cef1fc974e16da173fdf829d