Refactor WPTs to support manual reportEvent beacons.
We have a few different workstreams converging on manual
reportEvent beacon changes, but we can't really test
for them in the WPT environment. This CL registers a
click URL in the bidder's registerAdBeacon, along with some
additional refactoring and renaming to indicate that tests
support more than just automatic beacons now.

Change-Id: Ibf35acafd335bc1d6e195574b64db6479ee0e7ec