Add animation support for sizing keywords as lengths.
This adds animation support for sizing keywords as lengths, and uses it
only when calc-size() is enabled.  It then forbids animating the values.
Future CLs will be needed to make some cases interpolate (but some cases
should not interpolate, and we don't have all the infrastructure to
distinguish them yet).

This also adds a lot of tests for what should eventually work.  The
failures of the new tests are either because the above interpolation
support hasn't been added, or because we don't support the stretch
keyword yet.

Bug: 313072
Change-Id: I7f00e423ca155bdceace224ae34041ce458d063d