Limit use of calc-size() within other expression contexts.
This restricts the use of calc-size() to the toplevel expression and to
the basis of another calc-size() expression.  (It should eventually also
be allowed in mix() when we support that.)  It implements the
restriction in the same way that the spec describes it: based on the
idea that its type is not a length.

Bug: 313072
Change-Id: I157a16fee67674511f22f7d604c259c2230ab79c