[anchor] Support try-tactic flips for self-alignment properties
The self-alignment properties are a bit different from the other
properties that require adjustment.

For insets and sizes, a value transformation (e.g. anchor(right)
becoming anchor(left)) is always accompanied by a corresponding property
transformation (e.g. left becoming right), but for self-alignment
properties we need value transformation within the same property.
E.g. justify-self:start under flip-start becomes justify-self:start
(no change in property). Therefore these properties are always present
in the try-tactics layer.

Also, it's worth noting that justify-self accepts the 'left' and
'right' keywords, but align-self does not. This means we can't
just naively grab-and-flip the value of the other property for
flip-start: we need to convert left/right to logical to avoid
giving align-self an invalid value. Since the exact details of
what left/right should become in this case is non-obvious
(and not specified), I've avoided this case in the WPT for now.

Bug: 40279608
Change-Id: I8fa675a0be3f3c50ca73b4b628583765e252b032